chg: :truck: rename _node submodule as node, _config as config
